# src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Confi API doc"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "tags": [
        {
            "name": "bookings"
        }
    ],
    "schemes": [
        "http"
    ],
    "paths": {
        "/bookings": {
            "put": {
                "tags": [
                    "bookings"
                ],
                "summary": "Create booking",
                "description": "Creates or updates existing booking and sends email when created",
                "operationId": "addBooking",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "booking object",
                        "required": true,
                        "schema": {
                            "properties": {
                                "firstName": {
                                    "type": "string"
                                },
                                "lastName": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "phoneNumber": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": false
                                },
                                "data": {
                                    "properties": {
                                        "booking": {
                                            "$ref": "#/definitions/Booking"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "x-err1": {
                        "description": "No first name",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": true
                                },
                                "msg": {
                                    "example": "firstName is required"
                                }
                            }
                        }
                    },
                    "x-err2": {
                        "description": "No last name",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": true
                                },
                                "msg": {
                                    "example": "lastName is required"
                                }
                            }
                        }
                    },
                    "x-err3": {
                        "description": "No email",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": true
                                },
                                "msg": {
                                    "example": "email is required"
                                }
                            }
                        }
                    },
                    "x-err4": {
                        "description": "No phone number",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": true
                                },
                                "msg": {
                                    "example": "phoneNumber is required"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "bookings"
                ],
                "summary": "Get booking list",
                "operationId": "getBookings",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "example": false
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "bookings": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/Booking"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/bookings/{id}": {
            "delete": {
                "tags": [
                    "bookings"
                ],
                "summary": "Delete booking by id",
                "operationId": "deleteBooking",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "booking id",
                        "required": true,
                        "type": "string",
                        "format": "ObjectId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "example": false
                                },
                                "data": {
                                    "example": {}
                                }
                            }
                        }
                    },
                    "x-err1": {
                        "description": "Invalid id",
                        "schema": {
                            "properties": {
                                "error": {
                                    "example": true
                                },
                                "msg": {
                                    "example": "Invalid id value"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        }
    },
    "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    },
    "definitions": {
        "Booking": {
            "type": "object",
            "required": [
                "firstName",
                "lastName",
                "email",
                "phoneNumber",
                "confirmationCode"
            ],
            "properties": {
                "_id": {
                    "type": "string",
                    "format": "ObjectId"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "confirmationCode": {
                    "type": "string"
                }
            }
        }
    }
}